requirements.txt
anyio==4.0.0
asyncio==3.4.3
certifi==2023.7.22
exceptiongroup==1.1.3
feedparser==6.0.10
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
